requirements.txt
pygame
moderngl
numpy
pyrr
gltflib
pillow